Adjust for spot grid

+++ internal/model/pair.go
@@ -2,11 +2,11 @@ package model
 
 type Pair struct {
 	BaseEntity
-	BaseID     int      `json:"baseId" validate:"required"`
+	BaseID     int64    `json:"baseId" validate:"required"`
 	Base       Currency `json:"base"`
-	QuoteID    int      `json:"quoteId" validate:"required"`
+	QuoteID    int64    `json:"quoteId" validate:"required"`
 	Quote      Currency `json:"quote"`
-	ExchangeID int      `json:"exchangeId" validate:"required"`
+	ExchangeID int64    `json:"exchangeId" validate:"required"`
 	Exchange   Exchange `json:"exchange"`
 	Symbol     string   `json:"symbol" validate:"required"`
 }
